Test: expressions and const
