+++ tests/const.rs
@@ -1,8 +1,9 @@
 use crate::utils::SemanticTest;
-use semantic_analyzer::ast::{self, GetName, MAX_PRIORITY_LEVEL_FOR_EXPRESSIONS};
+use semantic_analyzer::ast::{self, GetName};
 use semantic_analyzer::types::expression::ExpressionOperations;
 use semantic_analyzer::types::{
-    semantic::SemanticStackContext, Constant, ConstantName, ConstantValue, PrimitiveValue,
+    semantic::SemanticStackContext, Constant, ConstantExpression, ConstantName, ConstantValue,
+    PrimitiveValue,
 };
 
 mod utils;
@@ -55,68 +56,22 @@ fn const_value_ast_transform() {
 
 #[test]
 fn const_expr_ast_transform() {
-    let _cnt_expr = ast::ConstantExpression {
-        value: ast::ConstantValue::Value(ast::PrimitiveValue::I8(10)),
+    let cnt_expr_prev = ast::ConstantExpression {
+        value: ast::ConstantValue::Value(ast::PrimitiveValue::I16(20)),
         operation: None,
     };
-}
-
-#[test]
-fn ast_expression_operations() {
-    assert_eq!(ast::ExpressionOperations::Plus.priority(), 5);
-    assert_eq!(ast::ExpressionOperations::Minus.priority(), 4);
-    assert_eq!(ast::ExpressionOperations::Divide.priority(), 8);
-    assert_eq!(
-        ast::ExpressionOperations::Multiply.priority(),
-        MAX_PRIORITY_LEVEL_FOR_EXPRESSIONS
-    );
-    assert_eq!(
-        ast::ExpressionOperations::ShiftLeft.priority(),
-        MAX_PRIORITY_LEVEL_FOR_EXPRESSIONS
-    );
-    assert_eq!(
-        ast::ExpressionOperations::ShiftRight.priority(),
-        MAX_PRIORITY_LEVEL_FOR_EXPRESSIONS
-    );
-    assert_eq!(ast::ExpressionOperations::Or.priority(), 6);
-    assert_eq!(ast::ExpressionOperations::Xor.priority(), 6);
-    assert_eq!(ast::ExpressionOperations::And.priority(), 7);
-    assert_eq!(ast::ExpressionOperations::Eq.priority(), 7);
-    assert_eq!(ast::ExpressionOperations::NotEq.priority(), 7);
-    assert_eq!(ast::ExpressionOperations::Great.priority(), 7);
-    assert_eq!(ast::ExpressionOperations::GreatEq.priority(), 7);
-    assert_eq!(ast::ExpressionOperations::LessEq.priority(), 7);
-
-    let ex_op: ExpressionOperations = ast::ExpressionOperations::Plus.into();
-    assert_eq!(ex_op, ExpressionOperations::Plus);
-    let ex_op: ExpressionOperations = ast::ExpressionOperations::Minus.into();
-    assert_eq!(ex_op, ExpressionOperations::Minus);
-    let ex_op: ExpressionOperations = ast::ExpressionOperations::Multiply.into();
-    assert_eq!(ex_op, ExpressionOperations::Multiply);
-    let ex_op: ExpressionOperations = ast::ExpressionOperations::Divide.into();
-    assert_eq!(ex_op, ExpressionOperations::Divide);
-    let ex_op: ExpressionOperations = ast::ExpressionOperations::ShiftLeft.into();
-    assert_eq!(ex_op, ExpressionOperations::ShiftLeft);
-    let ex_op: ExpressionOperations = ast::ExpressionOperations::ShiftRight.into();
-    assert_eq!(ex_op, ExpressionOperations::ShiftRight);
-    let ex_op: ExpressionOperations = ast::ExpressionOperations::Or.into();
-    assert_eq!(ex_op, ExpressionOperations::Or);
-    let ex_op: ExpressionOperations = ast::ExpressionOperations::And.into();
-    assert_eq!(ex_op, ExpressionOperations::And);
-    let ex_op: ExpressionOperations = ast::ExpressionOperations::Xor.into();
-    assert_eq!(ex_op, ExpressionOperations::Xor);
-    let ex_op: ExpressionOperations = ast::ExpressionOperations::Eq.into();
-    assert_eq!(ex_op, ExpressionOperations::Eq);
-    let ex_op: ExpressionOperations = ast::ExpressionOperations::NotEq.into();
-    assert_eq!(ex_op, ExpressionOperations::NotEq);
-    let ex_op: ExpressionOperations = ast::ExpressionOperations::Great.into();
-    assert_eq!(ex_op, ExpressionOperations::Great);
-    let ex_op: ExpressionOperations = ast::ExpressionOperations::Less.into();
-    assert_eq!(ex_op, ExpressionOperations::Less);
-    let ex_op: ExpressionOperations = ast::ExpressionOperations::GreatEq.into();
-    assert_eq!(ex_op, ExpressionOperations::GreatEq);
-    let ex_op: ExpressionOperations = ast::ExpressionOperations::LessEq.into();
-    assert_eq!(ex_op, ExpressionOperations::LessEq);
+    let cnt_expr: ConstantExpression = ast::ConstantExpression {
+        value: ast::ConstantValue::Value(ast::PrimitiveValue::I8(10)),
+        operation: Some((ast::ExpressionOperations::Eq, Box::new(cnt_expr_prev))),
+    }
+    .into();
+    let c_val: ConstantValue = ast::ConstantValue::Value(ast::PrimitiveValue::I8(10)).into();
+    assert_eq!(cnt_expr.value, c_val);
+    let expr_op: ExpressionOperations = ast::ExpressionOperations::Eq.into();
+    assert_eq!(cnt_expr.operation.clone().unwrap().0, expr_op);
+    let c_val2: ConstantValue = ast::ConstantValue::Value(ast::PrimitiveValue::I16(20)).into();
+    assert_eq!(cnt_expr.operation.clone().unwrap().1.value, c_val2);
+    assert_eq!(cnt_expr.operation.unwrap().1.operation, None);
 }
 
 #[test]
@@ -129,6 +84,35 @@ fn const_declaration() {
         constant_value: ast::ConstantExpression {
             value: ast::ConstantValue::Value(ast::PrimitiveValue::I8(10)),
             operation: None,
+        },
+    };
+    t.state.constant(&const_statement);
+    assert!(t.state.global.constants.contains_key(&const_name.into()));
+    assert!(t.state.errors.is_empty());
+    let state = t.state.global.context.clone().get();
+    assert_eq!(state.len(), 1);
+    assert_eq!(
+        state[0],
+        SemanticStackContext::Constant {
+            const_decl: const_statement.clone().into()
+        }
+    );
+}
+
+#[test]
+fn const_declaration_with_operarions() {
+    let mut t = SemanticTest::new();
+    let cnt_expr_prev = ast::ConstantExpression {
+        value: ast::ConstantValue::Constant(ast::ConstantName::new(ast::Ident::new("cnt2"))),
+        operation: None,
+    };
+    let const_name = ast::ConstantName::new(ast::Ident::new("cnt1"));
+    let const_statement = ast::Constant {
+        name: const_name.clone(),
+        constant_type: ast::Type::Primitive(ast::PrimitiveTypes::I8),
+        constant_value: ast::ConstantExpression {
+            value: ast::ConstantValue::Value(ast::PrimitiveValue::I8(10)),
+            operation: Some((ast::ExpressionOperations::Plus, Box::new(cnt_expr_prev))),
         },
     };
     t.state.constant(&const_statement);
